requirements.txt
Django==5.2.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.10
requests==2.32.4
gunicorn==23.0.0
whitenoise==6.9.0
dj-database-url==2.2.0